corrected type hinting for the `timeout` argument (int -> float) to the `embed_confs()` function in `conformers.py`

+++ bullviso/conformers.py
@@ -173,7 +173,7 @@ def embed_confs(
     mol: Chem.Mol,
     n_confs: int = None,
     rmsd_threshold: float = 0.0,
-    timeout: int = None,
+    timeout: float = None,
     coord_map: dict[int, rdGeometry.Point3D] = None,
     n_proc: int = 1,
     seed: int = -1
@@ -197,8 +197,8 @@ def embed_confs(
         rmsd_threshold (float, optional): RMSD threshold for deduplicating
             embeddings; conformers with RMSDs below the RMSD threshold are
             considered to be duplicates. Defaults to 0.0 (Angstroem).
-        timeout (int, optional): Timeout (in seconds) for conformer embedding.
-            Defaults to `None`.
+        timeout (float, optional): Timeout (in seconds) for conformer
+            embedding. Defaults to `None`.
         coord_map (dict[int, rdGeometry.Point3D], optional): Coordinate map
             dictionary mapping atom indices to their 3D coordinates
             (represented as rdGeometry.Point3D instances); these atoms are
